requirements.txt
charset-normalizer==2.1.0
folium==0.12.1.post1
Jinja2==3.1.2
numpy==1.23.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtChart-Qt5==5.15.2
pyqtgraph==0.12.4
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
requests==2.28.1
urllib3==1.26.11
